RDKCOM-3438: REFPLTV-1550 ctrlm-main build failure in RPi4 Hybrid 64bit
Reason for change: Type Conversion warnings needed type casting to fix the issue
Test Procedure: Build should be successful. ctrl manager should build fine.
Risks: Low.

(cherry picked from commit faf7247ce8861b0ce2d0461251518ced8950a4fd)
(cherry picked from commit f06883be0bb19f435b18e4d2f9cc9b4e944e4c91)

+++ src/voice/ctrlm_voice_obj.cpp
@@ -1536,7 +1536,7 @@ void ctrlm_voice_t::voice_session_data_post_processing(int bytes_sent, const cha
     if(session_delta > 0 && watermark > voice_buffer_high_watermark) {
         voice_buffer_high_watermark = watermark;
     }
-    if(session_delta > (VOICE_BUFFER_WARNING_THRESHOLD * this->voice_packet_interval)) {
+    if(session_delta > (long long)(VOICE_BUFFER_WARNING_THRESHOLD * this->voice_packet_interval)) {
         voice_buffer_warning_triggered = 1;
     }
     #endif
